app/pyproject.toml
[tool.poetry]
name = "Genesis"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
django-imagekit = "^5.0.0"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.7.0"
flake8 = "^6.1.0"
django-debug-toolbar = "^4.2.0"
drf-yasg = "^1.21.7"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py311"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests/*.py"]